Extract wireless lab helpers from app
